Initial commits -- Likely very Broken!!!
May work, may not.  Gotta figure out the pin for the 2nd board and how to use the data.  This is my best first guess.
Added duplicate information for pulling amperages and voltages from solar enabled whole house board that uses 2 energy meter boards from CircuitSetup.US

This will be very broken and is acting as a test bed at this point.  This will be functional soon.